Moved test merkle tree to it's own file
