+++ crypto/src/merkle_tree/proof.rs
@@ -108,7 +108,7 @@ mod tests {
     use lambdaworks_math::traits::ByteConversion;
 
     use crate::merkle_tree::{
-        merkle::MerkleTree, proof::Proof, TestHasher, TestMerkleTreeEcgfp, TestProofEcgfp5, Ecgfp5FE,
+        merkle::MerkleTree, proof::Proof, test_merkle::{Ecgfp5FE, TestProofEcgfp5, TestHasher, TestMerkleTreeEcgfp}
     };
 
     use lambdaworks_math::field::{element::FieldElement, fields::u64_prime_field::U64PrimeField};
