+++ proving_system/stark/src/fri/fri_decommit.rs
@@ -1,11 +1,9 @@
 use crate::fri::fri_commitment::FriCommitmentVec;
 pub use lambdaworks_crypto::fiat_shamir::transcript::Transcript;
-use lambdaworks_crypto::merkle_tree::TestHasher;
-
 use lambdaworks_crypto::merkle_tree::proof::Proof;
 use lambdaworks_math::field::element::FieldElement;
 use lambdaworks_math::field::traits::IsField;
-
+use crate::fri::TestHasher;
 #[derive(Debug, Clone)]
 pub struct FriDecommitment<F: IsField> {
     pub layer_merkle_paths: Vec<(Proof<F, TestHasher>, Proof<F, TestHasher>)>,
